requirements.txt
azure-ai-documentintelligence
python-dotenv
openai